Refactoring v2
