+++ app/assets/javascripts/models/beta/column.js
@@ -12,7 +12,7 @@ export const isBacklog = (story, project) => {
   return !isChillyBin(story) && (!Story.isAccepted(story) || isFromCurrentSprint);
 };
 
-export const isDone = column => column !== DONE;
+export const isNotDone = column => column !== DONE;
 
 export const DONE = 'done';
 export const BACKLOG = 'backlog';
